requirements.txt
streamlit
easyocr
numpy
pillow
opencv-python-headless